% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_labs.R
\name{deploy_labs}
\alias{deploy_labs}
\title{Deploy all labs to shinyapps.io}
\usage{
deploy_labs(labvec = 1:13)
}
\arguments{
\item{labvec}{vector with numbers indicating which labs to deploy. default is all.}
}
\value{
nothing
}
\description{
This function sends each lab to shinyapps.io
}
\details{
You need to make sure you have your shinyapps.io account set up and connected
The \code{rsconnect} package needs to be installed.
}
