% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labs.R
\name{get_labs}
\alias{get_labs}
\title{Get a zip file with all labs for deployment}
\usage{
get_labs()
}
\value{
nothing
}
\description{
This function opens a file explorer
pointing the user to the location of zip file with all labs
}
