% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iblirmenu.R
\name{iblirmenu}
\alias{iblirmenu}
\title{The main menu for the iblir package}
\usage{
iblirmenu()
}
\description{
This function opens a menu that will allow the user to run the different simulations.
}
\details{
Run this function with no arguments to start the main menu
}
\examples{
\dontrun{iblirmenu()}
}
\author{
Andreas Handel
}
