% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_labs.R
\name{prepare_labs}
\alias{prepare_labs}
\title{Prepare all labs for use}
\usage{
prepare_labs(labfolder = NULL)
}
\arguments{
\item{labfolder}{optional string pointing to a folder in which the labs are located. if not provided, the location for the default labs is used.}
}
\value{
nothing
}
\description{
This function makes a zip folder of all labs and generates a landing page
}
\details{
This only needs to be run during development of the labs, or if you make changes.
}
