% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_labs.R
\name{deploy_labs}
\alias{deploy_labs}
\title{Deploy all labs to shinyapps.io}
\usage{
deploy_labs(labvec = 1:13, labfolder = NULL)
}
\arguments{
\item{labvec}{optional vector with numbers indicating which labs to deploy. Default is all. Ignored if \code{labfolder} is provided.}

\item{labfolder}{optional string pointing to a folder in which the labs are located. if not provided, the location for the default labs is used.}
}
\value{
nothing
}
\description{
This function sends each lab to shinyapps.io
}
\details{
You need to make sure you have your shinyapps.io account set up and connected.
The \code{rsconnect} package needs to be installed as well.
}
