% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labs.R
\name{get_labs}
\alias{get_labs}
\title{Get a zip file with all labs for deployment}
\usage{
get_labs()
}
\value{
Location of the zip file which the user can save.
}
\description{
This function gives the user a zip file with all labs
this can be deployed to a shiny server
}
