% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iblir.R
\docType{package}
\name{iblir}
\alias{iblir}
\title{iblir: A package to learn basic stats}
\description{
The package provides a number of shiny/learnr apps
To learn more about how to use the package, see the vignette
or the short introduction on the package github repository.
https://github.com/andreashandel/iblir
}
\section{Running a tutorial}{


To run a tutorial execute:

\code{learnr::run_tutorial("name_of_tutorial", package = "adventr")}

replacing \emph{name_of_tutorial} with the name of the tutorial (i.e., LAbN, whith N the lab number). For example, to run lab 2:

\code{learnr::run_tutorial("Lab2", package = "iblir")}
}

